Тестовый генератор
Компилятор Pixic близок к выпуску как никогда! Теперь вы можете протестировать ещё одну его важную составляющую: генератор промежуточного кода на языке C. Правда, пока это возможно сделать только с ограниченным множеством типов инструкций и выражений, но это уже что-то. Подробности по тестированию в файле "README.md" проекта.
Здесь используются возможности нового, недавно вышедшего Pixilang 3.8.3, а именно список свойств. Так что перед тестированием убедитесь, что он у вас установлен.
Из прочего я сделал логотип для Pixic: HD векторная версия существа с логотипа Pixilang, которое ещё быстро бежит, потому что Pixic быстрый. Ещё решил писать модульные тесты. Я, как бы, тесты всегда пишу, только я их потом удаляю. А тут я решил их хранить. Ну, и плюс это ещё что-то типа интерактивной документации.
Так, что ещё рассказать? Может, анекдот какой-нибудь? Не, ладно. Тестируйте, в общем.

+++ src/Pixic runtime/src/library/include/core/Container.h
@@ -50,8 +50,8 @@ struct Container
 		double *float64;
 		#endif
 	};
-	Int	width, height;
-	unsigned long flags;
+	Int width, height;
+	uint32_t flags;
 	Color key_color;
 	Container_ID alpha;
 };
